+++ docs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @midl-xyz/midl-js-docs
 
+## 0.0.17
+
+### Patch Changes
+
+- Updated dependencies
+  - @midl-xyz/midl-js-core@0.0.14
+  - @midl-xyz/midl-js-react@0.0.18
+
 ## 0.0.16
 
 ### Patch Changes