+++ playground/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @midl-xyz/midl-js-playground
 
+## 0.1.21
+
+### Patch Changes
+
+- fix: edict multiple runes
+- Updated dependencies
+  - @midl-xyz/midl-js-core@0.0.14
+  - @midl-xyz/midl-js-executor@0.0.20
+  - @midl-xyz/midl-js-react@0.0.18
+
 ## 0.1.20
 
 ### Patch Changes